Add RAG Build Tag Testing Scripts
- Introduced `test-rag-build-tags.ps1` for PowerShell users to test RAG client implementations with and without the 'rag' build tag, including detailed output for tests, benchmarks, race condition detection, and coverage analysis.
- Added `test-rag-build-tags.sh` for Bash users, mirroring the functionality of the PowerShell script with appropriate shell syntax and error handling.
- Created `test-rag-simple.sh` to perform targeted tests on specific files, ensuring compatibility and avoiding compilation issues across packages. This script includes interface tests, configuration tests, and performance benchmarks for both no-op and Weaviate implementations.
- Enhanced output formatting with color coding for better readability and user experience.
