+++ pkg/rag/client_weaviate_old.go
